Add and setup webpack
