+++ package.json
@@ -1,8 +1,9 @@
 {
   "name": "@akshay-nm/confetch",
-  "version": "0.8.0",
+  "version": "0.8.1",
   "description": "Configur-ed/able fetch",
-  "main": "./src/index.js",
+  "main": "dist/confetch.js",
+  "module": "src/index.js",
   "repository": "git@github.com:akshay-nm/confetch.git",
   "author": "Akshay Kumar <akshay.nm92@gmail.com>",
   "license": "MIT",
@@ -12,16 +13,22 @@
   },
   "scripts": {
     "test": "jest",
-    "docs": "jsdoc -c ./jsdoc.conf.json"
+    "docs": "jsdoc -c ./jsdoc.conf.json",
+    "build": "webpack"
   },
   "devDependencies": {
-    "jest": "^26.6.3"
-  },
-  "dependencies": {
     "@pixi/jsdoc-template": "^2.6.0",
-    "abort-controller": "^3.0.0",
     "jaguarjs-jsdoc": "^1.1.0",
+    "jest": "^26.6.3",
     "jsdoc": "^3.6.6",
+    "val-loader": "^3.1.0",
+    "webpack": "^5.30.0",
+    "webpack-cli": "^4.6.0",
+    "abort-controller": "^3.0.0",
     "node-fetch": "^2.6.1"
-  }
+  },
+  "bundledDependencies": [
+    "abort-controller",
+    "node-fetch"
+  ]
 }
